pyproject.toml
[project]
name = "fhir-omop"
version = "0.1.0"
description = "Required packages"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.1",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.48",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
